v2.5 Public Department

+++ frontend/src/components/department/TableDepartmentComponent.jsx
@@ -130,6 +130,11 @@ const TableDepartmentComponent = ({ search, rowsPerPage, currentPage, setCurrent
 
     /** Modificar departamento **/
     const handleModify = async (depItem) => {
+        if (depItem.id === 1) {
+            Swal.fire("Error", "Este departamento no se puede eliminar", "error");
+            return;
+        }
+
         await AddModifyDepartmentComponent({
             depItem,
             action: "modify",
@@ -149,6 +154,11 @@ const TableDepartmentComponent = ({ search, rowsPerPage, currentPage, setCurrent
     const handleDelete = async (depItem) => {
         try { await showCaptcha(); }
         catch { return; }
+
+        if (depItem.id === 1) {
+            Swal.fire("Error", "Este departamento no se puede eliminar", "error");
+            return;
+        }
 
         const result = await deleteDepartment(depItem.id);
         if (result.success) {
@@ -174,12 +184,15 @@ const TableDepartmentComponent = ({ search, rowsPerPage, currentPage, setCurrent
                 </thead>
                 <tbody>
                     {currentDepartments.map((depItem, idx) => {
-                        const canModify = ["ADMIN", "SUPERADMIN"].includes(currentUser.usertype);
+                        const canModifyLinks = ["ADMIN", "SUPERADMIN"].includes(currentUser.usertype);
+                        const canModifyorDelete = ["ADMIN", "SUPERADMIN"].includes(currentUser.usertype) && depItem.id !== 1;
                         const depLinks = depItem.links || [];
                         const depAvailableLinks = links.filter(l => !depLinks.some(dl => dl.id === l.id));
                         return (
                             <tr key={idx}>
-                                <td>{depItem.id}</td>
+                                <td style={{
+                                    color: depItem.id === 1 ? "#0d6efd" : undefined,
+                                    fontWeight: 500}}> {depItem.id} </td>
                                 <DepartmentNameToolTipComponent
                                     depItem={depItem}
                                     isSmallScreen_v0={isSmallScreen_v0}
@@ -191,7 +204,7 @@ const TableDepartmentComponent = ({ search, rowsPerPage, currentPage, setCurrent
                                             <ShowMoreBadgeComponent
                                                 currentUser={currentUser}
                                                 user={depItem}
-                                                canModify={canModify}
+                                                canModify={canModifyLinks}
                                                 objType="enlace"
                                                 userObjects={depLinks}
                                                 availableObjs={depAvailableLinks}
@@ -204,7 +217,7 @@ const TableDepartmentComponent = ({ search, rowsPerPage, currentPage, setCurrent
                                                     window.dispatchEvent(new Event("refresh-departments"));
                                                 }}
                                             />
-                                        ) : canModify ? (
+                                        ) : canModifyLinks ? (
                                             <AddBadgeComponent
                                                 availableObjs={depAvailableLinks}
                                                 objType="enlace"
@@ -216,7 +229,7 @@ const TableDepartmentComponent = ({ search, rowsPerPage, currentPage, setCurrent
                                         ) : null
                                     ) : (
                                         <>
-                                            {depLinks.slice(0, 3).map(l => canModify
+                                            {depLinks.slice(0, 3).map(l => canModifyLinks
                                                 ? <RemovableBadgeComponent key={l.id} objName={l.name} objType="enlace" onDelete={async () => {
                                                     await deleteLinkToDepartment(depItem.id, l.id);
                                                     window.dispatchEvent(new Event("refresh-departments"));
@@ -227,7 +240,7 @@ const TableDepartmentComponent = ({ search, rowsPerPage, currentPage, setCurrent
                                                 <ShowMoreBadgeComponent
                                                     currentUser={currentUser}
                                                     user={depItem}
-                                                    canModify={canModify}
+                                                    canModify={canModifyLinks}
                                                     objType="enlace"
                                                     userObjects={depLinks}
                                                     availableObjs={depAvailableLinks}
@@ -241,7 +254,7 @@ const TableDepartmentComponent = ({ search, rowsPerPage, currentPage, setCurrent
                                                     }}
                                                 />
                                             )}
-                                            {depLinks?.length <= 3 && canModify && (
+                                            {depLinks?.length <= 3 && canModifyLinks && (
                                                 <AddBadgeComponent availableObjs={depAvailableLinks} objType="enlace" onAdded={async l => {
                                                     await addLinkToDepartment(depItem.id, l.id);
                                                     window.dispatchEvent(new Event("refresh-departments"));
@@ -252,7 +265,7 @@ const TableDepartmentComponent = ({ search, rowsPerPage, currentPage, setCurrent
                                 </td>
                                 <td className="text-center">
                                     <div className="d-flex justify-content-center flex-wrap gap-1">
-                                        {canModify && (
+                                        {canModifyorDelete && (
                                             <>
                                                 <Button color="warning" size="sm"
                                                     style={{
